Remember selections in graph page - 1.0.38 🚀
If a person picks body weight,
then the next time they navigate to
graphs, we should see body weight
selected.

The only reason this wasn't being
persisted to begin with is Flutter
destroying old routes that aren't in
focus. Which is good for performance.
All I did was lift the state up into a
Provider instead of using localized
setStates. Not sure what the performance
implications are of this, but i imagine
they are similar to setState.