Notify to log breakfast lunch & dinner
It's toggled on/off with the "Notifications"
setting. Also you can mute per category because
they have different notification channels.

I made this setting default to
false for future users. Im not sure
if most people want to be nagged
like this so it will be off by default.

For current users who have it
turned on (probably) it will stay on.
It might default to false even for
existing users if they never tapped
the toggle off/on.